+++ addon/ng2/blueprints/ng2/files/__path__/typings.d.ts
@@ -1,4 +1,3 @@
-
 // Typings reference file, see links for more information
 // https://github.com/typings/typings
 // https://www.typescriptlang.org/docs/handbook/writing-declaration-files.html
@@ -10,7 +9,7 @@
 
 <% if(!isMobile) { %>
 declare var System: any;
-declare var module: { id: string };
+<% if(!isMobile) { %>declare var module: { id: string };<% } %>
 declare var require: any;
 <% } %>